+++ server/models/_index.js
@@ -12,4 +12,5 @@ module.exports = {
   ProjectTypes: require('./project.js').ProjectTypes,
   ProjectTypeDisplayFields: require('./project.js').ProjectTypeDisplayFields,
   ProjectPermissionRequests: require('./project.js').ProjectPermissionRequests,
+  NCISupplyDemandExclusionList: require('./analytics.js').NCISupplyDemandExclusionList
 }