+++ src/app/_shared/services/_index.ts
@@ -1,4 +1,4 @@
-export * from './api-data.service';
+export * from './auth.service';
 export * from './cache.service';
 export * from './click-tracking.service';
 export * from './cookies.service';