+++ src/app/app.module.ts
@@ -24,16 +24,19 @@ import { FilterPipe } from './_shared/pipes/filter.pipe';
 import { TitleCasePipe } from '@angular/common';
 
 // SERVICES
-import { ApiDataService, CacheService, ClickTrackingService, CookiesService,
+import { CacheService, ClickTrackingService, CookiesService,
   ToolsService, UserResolverService, WebsocketService } from './_shared/services/_index';
 import { AuthService } from './_shared/services/auth.service';
-import { AuthGuardService } from './_shared/guards/auth.guard';
 import { ApiDataAuthService, ApiDataClickTrackingService, ApiDataEmailService, ApiDataEmployeeService,
   ApiDataFteService, ApiDataJobTitleService, ApiDataMetaDataService, ApiDataOrgService, ApiDataPermissionService, ApiDataProjectService,
   ApiDataReportService, ApiDataDashboardService, ApiDataAnalyticsService, ApiDataBomService} from './_shared/services/api-data/_index';
 
 // GUARDS
+import { AuthGuardService } from './_shared/guards/auth.guard';
 import { UnsavedChangesGuard } from './_shared/guards/unsaved-changes.guard';
+import { BrowserGuard } from './_shared/guards/browser.guard';
+import { PermissionsGuard } from './_shared/guards/permissions.guard';
+import { FteEntryGuard } from './fte-entry/employee/fte-entry.guard';
 
 // CHARTS
 import 'hammerjs';
@@ -56,7 +59,7 @@ import { ProjectsSetupsComponent } from './setups/projects/projects.component';
 import { ProfileModalComponent } from './modals/profile-modal/profile-modal.component';
 import { ProjectsModalComponent } from './modals/projects-modal/projects-modal.component';
 import { TestComponent } from './test/test.component';
-import { OrgDropdownComponent } from './reports/employees/org-dropdown/org-dropdown.component';
+import { OrgDropdownComponent } from './reports/team-fte-summary/org-dropdown/org-dropdown.component';
 import { ProjectsEditModalComponent } from './modals/projects-edit-modal/projects-edit-modal.component';
 import { ProjectsCreateModalComponent } from './modals/projects-create-modal/projects-create-modal.component';
 import { ChartsModule } from 'ng2-charts';
@@ -147,7 +150,6 @@ import { BomViewerComponent } from './reports/bom-viewer/bom-viewer.component';
     TreeModule
   ],
   providers: [
-    ApiDataService,
     ApiDataAuthService,
     ApiDataClickTrackingService,
     ApiDataEmailService,
@@ -171,6 +173,9 @@ import { BomViewerComponent } from './reports/bom-viewer/bom-viewer.component';
     WebsocketService,
     CookiesService,
     UnsavedChangesGuard,
+    BrowserGuard,
+    PermissionsGuard,
+    FteEntryGuard,
     TitleCasePipe
   ],
   bootstrap: [AppComponent]