+++ src/app/modals/projects-modal/projects-modal.component.ts
@@ -606,6 +606,8 @@ export class ProjectsModalComponent implements OnInit, AfterViewInit {
       }
       updateModalSubscription.unsubscribe();
     });
+
+    this.onRequestedProject(project);
   }